Big Update

+++ src/main/java/com/c2se/roomily/service/impl/ChatMessageServiceImpl.java
@@ -2,6 +2,7 @@ package com.c2se.roomily.service.impl;
 
 import com.c2se.roomily.config.StorageConfig;
 import com.c2se.roomily.entity.ChatMessage;
+import com.c2se.roomily.entity.ChatRoom;
 import com.c2se.roomily.entity.User;
 import com.c2se.roomily.enums.ErrorCode;
 import com.c2se.roomily.exception.APIException;
@@ -9,17 +10,16 @@ import com.c2se.roomily.exception.ResourceNotFoundException;
 import com.c2se.roomily.payload.request.ChatMessageToAdd;
 import com.c2se.roomily.payload.response.ChatMessageResponse;
 import com.c2se.roomily.repository.ChatMessageRepository;
-import com.c2se.roomily.repository.UserRepository;
+import com.c2se.roomily.repository.ChatRoomRepository;
 import com.c2se.roomily.service.ChatMessageService;
+import com.c2se.roomily.service.ChatRoomService;
 import com.c2se.roomily.service.StorageService;
+import com.c2se.roomily.service.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.time.LocalDateTime;
-import java.util.Comparator;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -28,74 +28,65 @@ import java.util.stream.Collectors;
 @AllArgsConstructor
 public class ChatMessageServiceImpl implements ChatMessageService {
     ChatMessageRepository chatMessageRepository;
-    UserRepository userRepository;
+    UserService userService;
+    ChatRoomRepository chatRoomRepository;
+    ChatRoomService chatRoomService;
     StorageService storageService;
     StorageConfig storageConfig;
     SimpMessagingTemplate messagingTemplate;
 
     @Override
     public ChatMessageResponse saveChatMessage(ChatMessageToAdd chatMessageToAdd) {
-        User sender = userRepository.findById(chatMessageToAdd.getSenderId())
-                .orElseThrow(() -> new ResourceNotFoundException("User", "id", chatMessageToAdd.getSenderId()));
-        User recipient = userRepository.findById(chatMessageToAdd.getRecipientId())
-                .orElseThrow(() -> new ResourceNotFoundException("User", "id", chatMessageToAdd.getRecipientId()));
-        String roomId = getChatRoomId(chatMessageToAdd.getSenderId(), chatMessageToAdd.getRecipientId());
-
+        User sender = userService.getUserEntity(chatMessageToAdd.getSenderId());
+        String roomId = chatMessageToAdd.getRoomId();
+        if (!chatRoomRepository.existsById(roomId))
+            throw new ResourceNotFoundException("Chat room", "id", roomId);
+        ChatRoom chatRoom = chatRoomRepository.findByIdLocked(roomId).orElseThrow(
+                () -> new ResourceNotFoundException("Chat room", "id", roomId)
+        );
         ChatMessage chatMessage = ChatMessage.builder()
                 .sender(sender)
                 .imageUrl(null)
-                .recipient(recipient)
                 .message(chatMessageToAdd.getContent())
                 .roomId(roomId)
+                .subId(chatRoom.getNextSubId()+1)
                 .build();
-
         if (chatMessageToAdd.getImage() != null) {
-            String fileName = roomId+ "_"+ UUID.randomUUID();
+            String fileName = roomId + "_" + UUID.randomUUID();
             try {
-                storageService.putObject(chatMessageToAdd.getImage(), storageConfig.getBucketStore(),fileName);
+                storageService.putObject(chatMessageToAdd.getImage(), storageConfig.getBucketStore(), fileName);
                 chatMessage.setImageUrl(storageService.generatePresignedUrl(storageConfig.getBucketStore(), fileName));
             } catch (Exception e) {
-                throw new APIException(HttpStatus.INTERNAL_SERVER_ERROR,
-                        ErrorCode.FLEXIBLE_ERROR,
+                throw new APIException(HttpStatus.INTERNAL_SERVER_ERROR, ErrorCode.FLEXIBLE_ERROR,
                         "Error while saving image");
             }
         }
-
+        chatRoom.setLastMessage(chatMessage.getMessage());
+        chatRoom.setLastMessageTimeStamp(chatMessage.getCreatedAt());
+        chatRoom.setLastMessageSender(chatMessage.getSender().getId());
+        chatRoom.setNextSubId(chatRoom.getNextSubId()+1);
+        chatRoomRepository.save(chatRoom);
         chatMessageRepository.save(chatMessage);
-        messagingTemplate.convertAndSendToUser(chatMessage.getRecipient().getId(),
-                "/queue/messages",
-                chatMessage);
+        List<String> users = chatRoomService.getChatRoomUserIds(roomId);
+        users.forEach(user -> messagingTemplate.convertAndSendToUser(user, "/queue/messages", chatMessage));
         return mapToResponse(chatMessage);
     }
 
     @Override
-    public List<ChatMessageResponse> getChatMessages(String user1,
-                                                     String user2,
-                                                     String pivot, String timestamp,
+    public List<ChatMessageResponse> getChatMessages(String userId, String roomId, String pivot, String timestamp,
                                                      int prev) {
-        String roomId = getChatRoomId(user1, user2);
+        String[] users = roomId.split("_");
+        if (!users[1].equals(userId) && !users[2].equals(userId))
+            throw new APIException(HttpStatus.BAD_REQUEST, ErrorCode.FLEXIBLE_ERROR, "Invalid room id");
         List<ChatMessage> chatMessages;
-        if (pivot == null && timestamp == null)
-            chatMessages = chatMessageRepository.findLastedByRoomId(roomId, prev);
-        else
-            chatMessages = chatMessageRepository.findByRoomId(roomId, pivot, timestamp, prev);
+        if (pivot == null && timestamp == null) chatMessages = chatMessageRepository.findLastedByRoomId(roomId, prev);
+        else chatMessages = chatMessageRepository.findByRoomId(roomId, pivot, timestamp, prev);
         return chatMessages.stream().map(this::mapToResponse).collect(Collectors.toList());
     }
 
     private ChatMessageResponse mapToResponse(ChatMessage chatMessage) {
-        return ChatMessageResponse.builder()
-                .id(chatMessage.getId())
-                .senderId(chatMessage.getSender().getId())
-                .recipientId(chatMessage.getRecipient().getId())
-                .message(chatMessage.getMessage())
-                .createdAt(chatMessage.getCreatedAt().toString())
-                .isRead(chatMessage.isRead())
-                .imageUrl(chatMessage.getImageUrl())
-                .build();
-    }
-
-    private String getChatRoomId(String user_1, String user_2) {
-        int compare = user_1.compareTo(user_2);
-        return compare < 0 ? user_1 + "_" + user_2 : user_2 + "_" + user_1;
+        return ChatMessageResponse.builder().id(chatMessage.getId()).senderId(chatMessage.getSender().getId()).message(
+                chatMessage.getMessage()).createdAt(chatMessage.getCreatedAt()).isRead(chatMessage.isRead()).imageUrl(
+                chatMessage.getImageUrl()).build();
     }
 }
